Expanded row for EDA role
