+++ frontend/eda/access/roles/RoleDetails.tsx
@@ -1,11 +1,3 @@
-import {
-  DescriptionList,
-  DescriptionListDescription,
-  DescriptionListGroup,
-  DescriptionListTerm,
-  Label,
-  LabelGroup,
-} from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import {
@@ -22,6 +14,7 @@ import { SWR_REFRESH_INTERVAL } from '../../common/eda-constants';
 import { edaAPI } from '../../common/eda-utils';
 import { EdaRole } from '../../interfaces/EdaRole';
 import { EdaRoute } from '../../main/EdaRoutes';
+import { EdaRolePermissions } from './components/EdaRolePermissions';
 
 export function RoleDetails() {
   const { t } = useTranslation();
@@ -29,22 +22,6 @@ export function RoleDetails() {
   const { data: role } = useGet<EdaRole>(edaAPI`/roles/${params.id ?? ''}/`, undefined, {
     refreshInterval: SWR_REFRESH_INTERVAL,
   });
-  const ResourceTypes = {
-    activation: t('Activation'),
-    activation_instance: t('Activation Instance'),
-    audit_rule: t('Audit Rule'),
-    audit_event: t('Audit Event'),
-    task: t('Task'),
-    user: t('User'),
-    project: t('Project'),
-    inventory: t('Inventory'),
-    extra_var: t('Extra Vars'),
-    playbook: t('Playbook'),
-    rulebook: t('Rulebook'),
-    role: t('Role'),
-    decision_environment: t('Decision environment'),
-    credential: t('Credential'),
-  };
 
   const getPageUrl = useGetPageUrl();
 
@@ -64,32 +41,7 @@ export function RoleDetails() {
         </PageDetails>
         <PageDetails numberOfColumns={'single'}>
           <PageDetail label={t('Permissions')}>
-            <DescriptionList
-              isCompact
-              isHorizontal
-              horizontalTermWidthModifier={{
-                default: '16ch',
-              }}
-            >
-              {role?.permissions.map((permission) => (
-                <DescriptionListGroup key={permission?.resource_type}>
-                  <DescriptionListTerm
-                    data-cy={'permissions'}
-                    style={{ fontWeight: 'normal' }}
-                    key={permission?.resource_type}
-                  >
-                    {ResourceTypes[permission?.resource_type] || permission?.resource_type}
-                  </DescriptionListTerm>
-                  <DescriptionListDescription>
-                    {!!permission?.action.length && (
-                      <LabelGroup numLabels={5}>
-                        {permission?.action.map((action) => <Label key={action}>{action}</Label>)}
-                      </LabelGroup>
-                    )}
-                  </DescriptionListDescription>
-                </DescriptionListGroup>
-              ))}
-            </DescriptionList>
+            <EdaRolePermissions role={role} />
           </PageDetail>
         </PageDetails>
       </Scrollable>
